requirements.txt
fastapi==0.111.0
uvicorn
transformers
torch
rasterio
pillow
aiohttp
gunicorn
opencv-python-headless==4.10.0.84
matplotlib
accelerate==0.32.1
bitsandbytes==0.43.1

